Split out Data Central logic to its own file
