+++ package.json
@@ -38,10 +38,10 @@
     "jest-fast-check": "^1.0.2",
     "jose": "^4.6.0",
     "jsonwebtoken": "^8.5.1",
+    "oauth": "0.9.15",
     "passport": "^0.5.2",
     "passport-oauth2": "^1.6.0",
     "pouchdb": "^7.2.2",
-    "pouchdb-adapter-memory": "^7.2.2",
     "pouchdb-find": "^7.3.0",
     "supertest": "^6.1.6",
     "ts-jest": "^27.1.3",
