+++ src/authhelpers.ts
@@ -21,7 +21,6 @@
 
 /* eslint-disable node/no-extraneous-import */
 import OAuth2Strategy from 'passport-oauth2';
-import type {OAuth2} from 'oauth';
 
 import {UserProfileCallback, DoneFunction} from './types';
 
@@ -37,28 +36,4 @@ export class CleanOAuth2Strategy extends OAuth2Strategy {
     }
     return this._userProfileHook(this._oauth2, accessToken, done);
   }
-}
-
-export function dc_auth_profile(
-  oauth2: OAuth2,
-  accessToken: string,
-  done: DoneFunction
-) {
-  oauth2.get(
-    'https://auth.datacentral.org.au/cas/oauth2.0/profile',
-    accessToken,
-    (err: any, body: any, res: any) => {
-      if (err) {
-        console.error('DC oauth profile error', err, body, res);
-        return done(err, err);
-      }
-
-      try {
-        const json = JSON.parse(body);
-        return done(null, json);
-      } catch (ex) {
-        return done(new Error('Failed to parse user profile'));
-      }
-    }
-  );
 }
